requirements.txt
# core dependencies
pandas>=1.3.0
openpyxl>=3.0.9
numpy>=1.20.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
fastapi>=0.95.0
uvicorn>=0.22.0
pyyaml>=6.0
pydantic>=1.10.0
loguru>=0.6.0

# optional dependencies
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
